--- libs/scir/src/netlist.rs
@@ -1,1 +1,0 @@
-
