+++ libs/spice/Cargo.toml
@@ -14,3 +14,4 @@ rust_decimal_macros = "1"
 
 scir = { version = "0.6.0", registry = "substrate", path = "../scir" }
 substrate = { version = "0.7.0", registry = "substrate", path = "../../substrate" }
+enumify = { version = "0.1.0", path = "../enumify", registry = "substrate" }
