+++ tests/src/hard_macro.rs
@@ -1,11 +1,11 @@
 use crate::paths::test_data;
 use crate::shared::buffer::BufferIo;
 
-use scir::netlist::{NetlistKind, NetlisterInstance};
 use serde::{Deserialize, Serialize};
 use sky130pdk::Sky130Pdk;
 #[cfg(feature = "spectre")]
 use spectre::Spectre;
+use spice::netlist::NetlisterInstance;
 use spice::Spice;
 use substrate::block::Block;
 use substrate::io::SchematicType;
@@ -111,8 +111,9 @@ fn export_hard_macro() {
     let spice_lib = lib.scir.convert_schema::<Spice>().unwrap().build().unwrap();
 
     let mut buf: Vec<u8> = Vec::new();
-    let netlister = NetlisterInstance::new(NetlistKind::Cells, &Spice, &spice_lib, &[], &mut buf);
-    netlister.export().unwrap();
+    Spice
+        .write_scir_netlist(&spice_lib, &mut buf, Default::default())
+        .unwrap();
     let string = String::from_utf8(buf).unwrap();
     println!("Netlist:\n{}", string);
 }
@@ -148,8 +149,7 @@ fn export_hard_macro_to_spectre() {
         .unwrap();
 
     let mut buf: Vec<u8> = Vec::new();
-    let netlister =
-        NetlisterInstance::new(NetlistKind::Cells, &Spectre {}, &spectre_lib, &[], &mut buf);
+    let netlister = NetlisterInstance::new(&Spectre {}, &spectre_lib, &mut buf, Default::default());
     netlister.export().unwrap();
     let string = String::from_utf8(buf).unwrap();
     println!("Netlist:\n{}", string);
@@ -170,8 +170,9 @@ fn export_inline_hard_macro() {
     let spice_lib = lib.scir.convert_schema::<Spice>().unwrap().build().unwrap();
 
     let mut buf: Vec<u8> = Vec::new();
-    let netlister = NetlisterInstance::new(NetlistKind::Cells, &Spice, &spice_lib, &[], &mut buf);
-    netlister.export().unwrap();
+    Spice
+        .write_scir_netlist(&spice_lib, &mut buf, Default::default())
+        .unwrap();
     let string = String::from_utf8(buf).unwrap();
     println!("Netlist:\n{}", string);
 }
